Release 1.0.0
First release of the pygame angle simulator, move the player with wasd and use the mouse to find a point on screen and find it's angle depending on the x-axis. The coordinate axis follows the player to assist in calculations, otherwise the angle and player positions wouldn't have lined up properly to be used.